Additional Login Errors fixed
Now upon logon if there is an error, login will be attempted 3 times. Generally when there is an error, its because the delay between inputs in the CLI is too short. Now the delay will increase by a second on each attempt, this solves the previously errors that we've seen.